Update README.md
Fetching dependencies need to be done before starting the services so I moved that up in the README. I also removed some basic git steps since this is common knowledge and add noise to the documentation. Finally, I also added the fetching of dependencies to the command that runs the integration tests.